backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart

python-dotenv
httpx
google-genai
playwright
pymupdf
python-docx
pytesseract
python-pptx
pandas
openpyxl
faiss-cpu
sentence-transformers
beautifulsoup4
langgraph
langchain-openai
langchain-core
pydantic>=2.0
python-telegram-bot
twilio
supabase
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
email-validator
python-multipart